Update DashboardViewport.ahk
Correct bug in which I was testing whether a qwerd's word existed, and in doing so adding an empty key to the qwerd dictionary. This caused some of my passwords to be added to the hash and to break the Editor and/or turn up in certain searches. Prepended a test for key exists before testing for word exists all over the codebase.